+++ conrad/__init__.py
@@ -2,13 +2,13 @@
 
 import os
 
+from click import get_app_dir
+
 from .__version__ import __version__
 
 
-USER_HOME = os.path.expanduser("~")
-CONRAD_HOME = os.path.join(USER_HOME, ".conrad")
-SQL_ALCHEMY_CONN = "sqlite:///{}/conrad.db".format(CONRAD_HOME)
-
+CONRAD_HOME = get_app_dir("conrad")
+SQL_ALCHEMY_CONN = f"sqlite:///{CONRAD_HOME}/conrad.db"
 
 if not os.path.exists(CONRAD_HOME):
     os.makedirs(CONRAD_HOME)