+++ crawlers/base.py
@@ -52,7 +52,7 @@ class BaseCrawler(object):
             v.validate(event)
             if v.errors:
                 for key, val in v.errors.items():
-                    print("{} - {}: {}".format(event["name"], key, val))
+                    print(f"{event['name']} - {key}: {val}")
 
         with open(filename, "w") as f:
             f.write(json.dumps(self.events, indent=4, sort_keys=True))